Added readonly sources

+++ src/main/java/hu/szeba/hades/model/task/data/SourceFile.java
@@ -14,8 +14,9 @@ public class SourceFile {
     private File file;
     private String name;
     private String data;
+    private boolean readonly;
 
-    public SourceFile(File file) throws IOException {
+    public SourceFile(File file, boolean readonly) throws IOException {
         this.file = file;
         this.name = file.getName();
         if (file.exists()) {
@@ -25,6 +26,7 @@ public class SourceFile {
         } else {
             this.data = "";
         }
+        this.readonly = readonly;
     }
 
     public String getName() { return name; }
@@ -34,17 +36,28 @@ public class SourceFile {
     }
 
     public void setData(String data) {
-        this.data = data;
+        if (!readonly) {
+            this.data = data;
+        }
     }
 
     public void save() throws IOException {
-        Files.write(Paths.get(file.getAbsolutePath()),
-                Arrays.stream(data.split("\n")).collect(Collectors.toList()));
+        if (!readonly) {
+            Files.write(Paths.get(file.getAbsolutePath()),
+                    Arrays.stream(data.split("\n")).collect(Collectors.toList()));
+        }
     }
 
     public void delete() throws IOException {
-        if (!file.delete()) {
+        if (readonly) {
+            throw new IOException("Source is readonly: " + name);
+        } else if (!file.delete()) {
             throw new IOException("Couldn't delete file: " + file.getAbsolutePath());
         }
     }
+
+    public boolean isReadonly() {
+        return readonly;
+    }
+
 }
